# docs/swagger.yaml
definitions:
...
  dto.CreateProjectRequest:
    properties:
      company:
        type: string
      contract_date:
        type: string
      contract_number:
        type: string
      description:
        type: string
      end_date:
        type: string
      name:
        type: string
      payment_method:
        type: string
      start_date:
        type: string
      status:
        type: string
      total_amount:
        type: number
      type:
        type: string
    required:
    - company
    - end_date
    - name
    - start_date
    - total_amount
    - type
    type: object
...
  models.Payment:
    properties:
      actual_date:
        description: 实际收款日期
        type: string
      amount:
        description: 金额
        type: number
      create_time:
        description: 创建时间
        type: string
      id:
        type: integer
      method:
        description: '收款方式 (如: 银行转账)'
        type: string
      percentage:
        description: 占总金额百分比
        type: number
      plan_date:
        description: 计划收款日期
        type: string
      project:
        allOf:
        - $ref: '#/definitions/models.Project'
        description: 关联
      project_id:
        description: 关联项目ID
        type: integer
      remark:
        description: 备注
        type: string
      stage:
        description: '款项阶段 (如: 首付款, 进度款, 尾款)'
        type: string
      status:
        description: '状态: uncollected, collected'
        type: string
      update_time:
        description: 更新时间
        type: string
      user_id:
        description: 经办人ID (通常为创建者或当前负责人)
        type: integer
    type: object
  models.Project:
    properties:
      company:
        description: 建设单位/客户
        type: string
      contract_date:
        description: 签订日期
        type: string
      contract_number:
        description: 合同编号
        type: string
      create_time:
        description: 创建时间
        type: string
      description:
        description: 项目描述
        type: string
      end_date:
        description: 计划结束日期
        type: string
      id:
        type: integer
      name:
        description: 项目名称
        type: string
      payment_method:
        description: 支付方式 (字典项)
        type: string
      payments:
        description: 关联款项列表
        items:
          $ref: '#/definitions/models.Payment'
        type: array
      received_amount:
        description: 已回款金额
        type: number
      start_date:
        description: 计划开始日期
        type: string
      status:
        description: '状态: pending, processing, completed, archived'
        type: string
      total_amount:
        description: 合同总金额
        type: number
      type:
        description: 项目类型 (字典项)
        type: string
      update_time:
        description: 更新时间
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联
      user_id:
        description: 负责人ID
        type: integer
    type: object
  models.User:
    properties:
      avatar:
        description: 头像 URL
        type: string
      create_time:
        description: 创建时间
        type: string
      department:
        description: 部门
        type: string
      email:
        description: 邮箱
        type: string
      id:
        type: integer
      last_login_time:
        description: 最后登录时间
        type: string
      name:
        description: 真实姓名
        type: string
      phone:
        description: 手机号
        type: string
      position:
        description: 职位
        type: string
      role:
        description: '角色: admin, user'
        type: string
      status:
        description: '状态: 1=正常, 0=禁用'
        type: integer
      update_time:
        description: 更新时间
        type: string
      username:
        description: 用户名，唯一
        type: string
    type: object
  response.PageData:
    properties:
      list:
        description: 当前页的数据列表
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  response.PageResult:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/response.PageData'
        description: 分页数据
      message:
        description: 提示消息
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/projects:
    get:
      description: 分页查询项目，支持按状态(status)和关键词(keyword)搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: '项目状态: pending, processing, completed, archived'
        in: query
        name: status
        type: string
      - description: '搜索关键词: 项目名称或合同编号'
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResult'
      security:
      - Bearer: []
      summary: 获取项目列表
      tags:
      - Project
    post:
      description: 创建一个新的项目记录
      parameters:
      - description: 项目信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
      security:
      - Bearer: []
      summary: 创建项目
      tags:
      - Project
  /api/v1/projects/{id}:
    delete:
      description: 软删除或硬删除项目记录(视Repo实现而定)，通常会级联删除关联款项
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
          schema:
            type: string
      security:
      - Bearer: []
      summary: 删除项目
      tags:
      - Project
    get:
      description: 根据项目ID获取详细信息
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
      security:
      - Bearer: []
      summary: 项目详情
      tags:
      - Project
    put:
      description: 更新项目的基础信息、状态或合同金额
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
      security:
      - Bearer: []
      summary: 更新项目
      tags:
      - Project
  /api/v1/projects/{id}/archive:
    post:
      description: 将项目标记为已归档，不再在常规列表中显示
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 归档成功
          schema:
            type: string
      security:
      - Bearer: []
      summary: 归档项目
      tags:
      - Project
...
  /api/v1/projects/check-contract-number:
    get:
      description: 检查输入的合同编号是否已被其他项目使用
      parameters:
      - description: 待检查的合同编号
        in: query
        name: contract_number
        required: true
        type: string
      - description: 排除的项目ID(编辑时使用)
        in: query
        name: exclude_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - Bearer: []
      summary: 检查合同号唯一性
      tags:
      - Project
  /api/v1/projects/generate-contract-number:
    get:
      description: 根据日期生成格式为 HTYYYYMMDDXXXX 的智能建议编号
      parameters:
      - description: 项目日期 (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 生成合同编号
      tags:
      - Project
...
swagger: "2.0"